+++ src/Chapter02.Tests/Listing02.09.CaseInsensitiveComparisonOfTwoStrings.Tests.cs
@@ -1,0 +1,15 @@
+
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_09.Tests;
+
+[TestClass]
+public class Comparison
+{
+    [TestMethod]
+    public void Main_WriteBooleanStatements()
+    {
+        const string expected = "Help Requested: True";
+
+        IntelliTect.TestTools.Console.ConsoleAssert.Expect(
+            expected, Program.Main);
+    }
+}
