+++ src/Chapter02.Tests/Listing02.10.DisplayingASingleQuoteUsingAnEscapeSequence.Tests.cs
@@ -1,15 +1,16 @@
-﻿
+
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_10.Tests;
 
 [TestClass]
-public class Comparison
+public class SingleQuoteTests
 {
     [TestMethod]
-    public void Main_WriteBooleanStatements()
+    public void Main_WriteSingleQuote()
     {
-        const string expected = "Help Requested: True";
+        const string expected =
+            @"'";
 
         IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-            expected, Program.Main);
+            expected, SingleQuote.Main);
     }
 }
