+++ src/Chapter02.Tests/Listing02.11.UsingUnicodeEncodingToDisplayASmileyFace.Tests.cs
@@ -5,10 +5,10 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_11.Tests;
 public class SingleQuoteTests
 {
     [TestMethod]
-    public void Main_WriteSingleQuote()
+    public void Main_WriteSmiley()
     {
         const string expected =
-            @"'";
+@":)";
 
         IntelliTect.TestTools.Console.ConsoleAssert.Expect(
             expected, SingleQuote.Main);
