--- src/Chapter02.Tests/Listing02.12.Tests.cs
@@ -1,16 +1,0 @@
-﻿
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_12.Tests;
-
-[TestClass]
-public class SingleQuoteTests
-{
-    [TestMethod]
-    public void Main_WriteSmiley()
-    {
-        const string expected =
-@":)";
-
-        IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-            expected, SingleQuote.Main);
-    }
-}
