+++ src/Chapter02.Tests/Listing02.12.UsingTheNewlineCharacterToInsertANewline.Tests.cs
@@ -1,6 +1,6 @@
-﻿using IntelliTect.TestTools.Console;
+using IntelliTect.TestTools.Console;
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_13.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_12.Tests;
 
 [TestClass]
 public class DuelOfWitsTests
