+++ src/Chapter02.Tests/Listing02.14.StringInterpolation.Tests.cs
@@ -1,5 +1,5 @@
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_15.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_14.Tests;
 
 [TestClass]
 public class InterpolationTests
