+++ src/Chapter02.Tests/Listing02.15.NewLinesWithinStringInterpolation.Tests.cs
@@ -1,9 +1,9 @@
 using Microsoft.CodeAnalysis;
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_17.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_15.Tests;
 
 [TestClass]
-public class CombiningLiteralsWithInterpolationTests
+public class NewLineInterpolationTests
 {
     [TestMethod]
     public void Main_CorrectOutput()
@@ -15,10 +15,12 @@ public class CombiningLiteralsWithInterpolationTests
 <<{firstName}
 >>Enter your last name: 
 <<{lastName}
->>Your full name is:
-{firstName} {lastName}";
+>>Your full name is: {firstName} {lastName}";
+
 
         IntelliTect.TestTools.Console.ConsoleAssert.Expect(expected,
               Program.Main);
+
+
     }
 }
