+++ src/Chapter02.Tests/Listing02.18.RawStringLiteralWithInterpolation.Tests.cs
@@ -1,5 +1,5 @@
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_19.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_18.Tests;
 
 [TestClass]
 public class RawLiteralsWithInterpolationTests
