+++ src/Chapter02.Tests/Listing02.19.DisplayingMobiusStripUsingRawStringLiteral.Tests.cs
@@ -1,5 +1,5 @@
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_20.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_19.Tests;
 
 [TestClass]
 public class TriangleTests
