+++ src/Chapter02.Tests/Listing02.20.UsingStringFormat.Tests.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_09.Tests;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_20.Tests;
 
 [TestClass]
 public class ProgramTests
