+++ src/Chapter02/Listing02.01.SpecifyingLiteralValues.cs
@@ -1,4 +1,4 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_01;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_01;
 
 public class Program
 {
