+++ src/Chapter02/Listing02.03.SpecifyingALiteralDecimal.cs
