+++ src/Chapter02/Listing02.05.ExponentialNotation.cs
