+++ src/Chapter02/Listing02.06.HexadecimalLiteralValue.cs
