+++ src/Chapter02/Listing02.07.BinaryLiteralValue.cs
