+++ src/Chapter02/Listing02.10.DisplayingASingleQuoteUsingAnEscapeSequence.cs
@@ -1,0 +1,12 @@
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_10;
+
+#region INCLUDE
+public class SingleQuote
+{
+    public static void Main()
+    {
+        Console.WriteLine('\'');
+    }
+}
+#endregion INCLUDE
+
