+++ src/Chapter02/Listing02.12.UsingTheNewlineCharacterToInsertANewline.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_13;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_12;
 
 public class DuelOfWits
 {
