--- src/Chapter02/Listing02.12.UsingUnicodeEncodingToDisplayASmileyFace.cs
@@ -1,13 +1,0 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_12;
-
-public class SingleQuote
-{
-    public static void Main()
-    {
-        #region INCLUDE
-        Console.Write('\u003A');
-
-        Console.WriteLine('\u0029');
-        #endregion INCLUDE
-    }
-}
