+++ src/Chapter02/Listing02.13.DisplayingATriangleUsingAVerbatimStringLiteral.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_14;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_13;
 
 public class Triangle
 {
