+++ src/Chapter02/Listing02.15.NewLinesWithinStringInterpolation.cs
@@ -1,5 +1,4 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_15;
-
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_15;
 public class Program
 {
     public static void Main()
@@ -15,7 +14,11 @@ public class Program
         lastName = Console.ReadLine();
 
         #region INCLUDE
-        Console.WriteLine($"Your full name is {firstName} {lastName}.");
+        Console.WriteLine($"Your full name is: { 
+            firstName +
+            " " +
+            lastName
+            }");
         #endregion INCLUDE
     }
 }
