+++ src/Chapter02/Listing02.16.CombiningStringLiteralsWithInterpolation.cs
@@ -1,4 +1,4 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_17;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_16;
 
 public class Program
 {
