+++ src/Chapter02/Listing02.18.RawStringLiteralWithInterpolation.cs
@@ -1,10 +1,11 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_16;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_18;
 public class Program
 {
     public static void Main()
     {
         string firstName;
         string lastName;
+
 
         Console.WriteLine("Enter your first name: ");
         firstName = Console.ReadLine();
@@ -14,11 +15,8 @@ public class Program
         lastName = Console.ReadLine();
 
         #region INCLUDE
-        Console.WriteLine($"Your full name is: { 
-            firstName +
-            " " +
-            lastName
-            }");
+        Console.WriteLine(
+            $"""Hello, I'm {firstName}. {firstName} {lastName}""");
         #endregion INCLUDE
     }
 }
