+++ src/Chapter02/Listing02.19.DisplayingMobiusStripUsingRawStringLiteral.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_20;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_19;
 public class Triangle
 {
     public static void Main()
