+++ src/Chapter02/Listing02.22.UsingStringsLengthMember.cs
