+++ src/Chapter02/Listing02.24.WorkingWithStrings.cs
